+++ packages/create-yoshi-app/templates/flow-editor/typescript/src/types/external-types.d.ts
@@ -1,7 +1,6 @@
 // import { IWixStatic } from './wix-sdk';
 
 declare module '@wix/santa-site-renderer-testkit';
-declare module 'velocity';
 declare module '@wix/wix-experiments/dist/src/laboratory-testkit';
 declare module '@wix/wix-base-ui';
 declare module '@wix/wix-config-emitter';
