+++ packages/create-yoshi-app/tsconfig.json
@@ -4,8 +4,5 @@
     "rootDir": "src",
     "outDir": "build"
   },
-  "include": [
-    "src/**/*",
-    "src/**/*.json"
-  ]
+  "include": ["src/**/*", "src/**/*.json"]
 }
