+++ packages/yoshi-flow-editor/package.json
@@ -15,6 +15,7 @@
   "dependencies": {
     "arg": "^4.1.1",
     "bfj": "^7.0.1",
+    "body-parser": "^1.19.0",
     "chalk": "^2.4.2",
     "fs-extra": "^8.1.0",
     "globby": "^10.0.1",
@@ -28,9 +29,15 @@
   },
   "devDependencies": {
     "@babel/core": "^7.7.4",
+    "@types/cors": "^2.8.6",
+    "@types/express": "^4.17.2",
     "@types/fs-extra": "^8.0.1",
-    "@types/react-dev-utils": "^9.0.1",
     "@types/webpack": "^4.39.8",
-    "@types/webpack-node-externals": "^1.6.3"
+    "@types/webpack-node-externals": "^1.6.3",
+    "@wix/wix-bootstrap-testkit": "^1.1.3955",
+    "@wix/wix-config-emitter": "^1.0.1395",
+    "@wix/wix-http-testkit": "^1.0.1318",
+    "cors": "^2.8.5",
+    "velocity": "^0.7.2"
   }
 }
