+++ packages/yoshi-flow-editor/src/external-types.d.ts
@@ -1,5 +1,9 @@
 declare module 'bfj';
 declare module 'yoshi-flow-legacy/bin/yoshi-legacy';
-declare module 'react-dev-utils/ModuleNotFoundPlugin';
 declare module 'rtlcss-webpack-plugin';
 declare module '@stylable/webpack-plugin';
+declare module '@wix/santa-site-renderer-testkit';
+declare module 'velocity';
+declare module '@wix/wix-experiments/dist/src/laboratory-testkit';
+declare module '@wix/wix-base-ui';
+declare module '@wix/wix-config-emitter';
