+++ packages/yoshi-flow-editor/src/server/vm.ts
@@ -1,7 +1,7 @@
 import { Engine } from 'velocity';
 
-import velocityData from '../velocity.data.json';
-import velocityDataPrivate from '../velocity.private.data.json';
+import velocityData from './velocity.data.json';
+import velocityDataPrivate from './velocity.private.data.json';
 
 export default (url: string, data = {}) =>
   new Engine({ template: url }).render({
